+++ lisp/cov2.lisp
@@ -1,7 +1,3 @@
-;;;; Sample coverage file
-
-(("/Users/ohler/go/src/github.com/ohler55/slip/test/testdata/cov-test.lisp" "00c69412bf934617158b59331dbca964db71d7aec8e2f28bb553e48d21239ec6"
-  "/Users/ohler/go/src/github.com/ohler55/slip/test/testdata/cov-test2.lisp" "55458ffb66e84d015f9e034214cbdada1a8f4d6024a1c8904d1306100f92db69")
- ("/Users/ohler/go/src/github.com/ohler55/slip/test/testdata/cov-test2.lisp" 1 1 1 14 1)
- ("/Users/ohler/go/src/github.com/ohler55/slip/test/testdata/cov-test.lisp" 1 1 2 11 1)
- ("/Users/ohler/go/src/github.com/ohler55/slip/test/testdata/cov-test.lisp" 2 4 2 10 0))
+(("/Users/ohler/go/src/github.com/ohler55/slip/lisp/test2.lisp" "9dfbd465e749723a7365f41897db9bc98a735605a58e97d8a66814c888cf4e7f")
+ ("/Users/ohler/go/src/github.com/ohler55/slip/lisp/test2.lisp" 1 1 2 11 1)
+ ("/Users/ohler/go/src/github.com/ohler55/slip/lisp/test2.lisp" 2 4 2 10 1))