+++ lisp/cov3.lisp
@@ -1,4 +1,4 @@
-(("/Users/ohler/.slip/config.lisp" "3a9036a097d027e28c13f170a0bd6cdc158aea9abec7c64ef59e1001e5e8efeb"
-  "/Users/ohler/go/src/github.com/ohler55/slip/lisp/test3.lisp" "7db6c3cc11321d2c5d792a4ba22e4fa64ec9ddd8fb680b3b98a4631037e150fb")
- ("/Users/ohler/go/src/github.com/ohler55/slip/lisp/test3.lisp" 3 1 3 13 1)
- ("/Users/ohler/go/src/github.com/ohler55/slip/lisp/test3.lisp" 3 6 3 12 1))
+(("/Users/ohler/go/src/github.com/ohler55/slip/lisp/test3.lisp" "30f5b4c0d003984da1a6cd69442e0502e751b556979eb708dc14356544ee85f1")
+ ("/Users/ohler/go/src/github.com/ohler55/slip/lisp/test3.lisp" 2 1 4 11 1)
+ ("/Users/ohler/go/src/github.com/ohler55/slip/lisp/test3.lisp" 3 4 4 10 1)
+ ("/Users/ohler/go/src/github.com/ohler55/slip/lisp/test3.lisp" 3 7 3 13 1))